Working on Final Project
Trying to see If I can work from the server given code to get to what we had in the mini project. Finding it harder to reset some info and might need to use the SQL stuff to help reduce that. Might want to add in the function of keeping the port open until the player exits the game.